requirements.txt
alembic==1.7.5
anyio==3.3.4
asgiref==3.4.1
botocore==1.23.37
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.72.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
minio==7.1.2
orjson==3.6.5
passlib==1.7.4
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.17.1
typing-extensions==3.10.0.2
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
